Extract interface from FailedDispatchQueryFactory
