+++ presto-main/src/main/java/com/facebook/presto/dispatcher/FailedDispatchQueryFactory.java
@@ -14,50 +14,25 @@
 package com.facebook.presto.dispatcher;
 
 import com.facebook.presto.Session;
-import com.facebook.presto.event.QueryMonitor;
-import com.facebook.presto.execution.ExecutionFailureInfo;
-import com.facebook.presto.execution.LocationFactory;
-import com.facebook.presto.server.BasicQueryInfo;
 import com.facebook.presto.spi.resourceGroups.ResourceGroupId;
 
-import javax.inject.Inject;
-
 import java.util.Optional;
-import java.util.concurrent.ExecutorService;
 
-import static com.facebook.presto.util.Failures.toFailure;
-import static java.util.Objects.requireNonNull;
-
-public class FailedDispatchQueryFactory
+/**
+ * Factory interface to create FailedDispatchQuery
+ *
+ * This interface is required for https://github.com/prestodb/presto/issues/23455
+ */
+public interface FailedDispatchQueryFactory
 {
-    private final QueryMonitor queryMonitor;
-    private final LocationFactory locationFactory;
-    private final ExecutorService executor;
-
-    @Inject
-    public FailedDispatchQueryFactory(QueryMonitor queryMonitor, LocationFactory locationFactory, DispatchExecutor dispatchExecutor)
-    {
-        this.queryMonitor = requireNonNull(queryMonitor, "queryMonitor is null");
-        this.locationFactory = requireNonNull(locationFactory, "locationFactory is null");
-        this.executor = requireNonNull(dispatchExecutor, "dispatchExecutor is null").getExecutor();
-    }
-
-    public FailedDispatchQuery createFailedDispatchQuery(Session session, String query, Optional<ResourceGroupId> resourceGroup, Throwable throwable)
-    {
-        ExecutionFailureInfo failure = toFailure(throwable);
-        FailedDispatchQuery failedDispatchQuery = new FailedDispatchQuery(
-                session,
-                query,
-                locationFactory.createQueryLocation(session.getQueryId()),
-                resourceGroup,
-                failure,
-                executor);
-
-        BasicQueryInfo queryInfo = failedDispatchQuery.getBasicQueryInfo();
-
-        queryMonitor.queryCreatedEvent(queryInfo);
-        queryMonitor.queryImmediateFailureEvent(queryInfo, failure);
-
-        return failedDispatchQuery;
-    }
+    /**
+     * If DispatchManager fails to create a DispatchQuery, a FailedDispatchQuery is created instead
+     *
+     * @param session session
+     * @param query query text
+     * @param resourceGroup resource group of the query
+     * @param throwable failure information about the query
+     * @return {@link FailedDispatchQuery}
+     */
+    FailedDispatchQuery createFailedDispatchQuery(Session session, String query, Optional<ResourceGroupId> resourceGroup, Throwable throwable);
 }
